Added ClientSettings for client-specific settings, e.g. different layout of video texture per-client.

+++ SimulCasterAudio/src/crossplatform/NetworkPipeline.cpp
@@ -56,7 +56,7 @@ namespace sca
 			stream.isDataLimitPerFrame = false;
 			stream.counter = 0;
 			stream.chunkSize = 2048;
-			stream.id = 100 + i;
+			stream.id = 100 + (uint32_t)i;
 			stream.dataType = avs::NetworkDataType::Audio;
 			streams.emplace_back(std::move(stream));
 		}
